microservices/serving/recommendation_api/requirements.txt
fastapi>=0.104,<1.0
uvicorn[standard]>=0.24
psycopg2-binary>=2.9
pyyaml>=6.0
python-jose[cryptography]>=3.3
